+++ lib/committee.rb
@@ -26,7 +26,6 @@ require_relative "committee/schema_validator/open_api_3/operation_object"
 require_relative "committee/schema_validator/open_api_3/request_validator"
 require_relative "committee/schema_validator/open_api_3/response_validator"
 require_relative "committee/schema_validator/open_api_3/string_params_coercer"
-require_relative "committee/schema_validator/open_api_3/string_path_params_coercer"
 require_relative "committee/schema_validator/hyper_schema"
 require_relative "committee/schema_validator/hyper_schema/request_validator"
 require_relative "committee/schema_validator/hyper_schema/response_generator"
