+++ lib/committee/schema_validator/open_api_3/operation_object.rb
@@ -10,18 +10,16 @@ module Committee
       @path_params = path_params
     end
 
-    def coerce_query_parameter_object(name, value)
-      query_parameter_object = query_parameters[name]
-      return [nil, false] unless query_parameter_object
-
-      coerce_value(value, query_parameter_object)
+    def coerce_path_parameter(query_hash, validator_option)
+      Committee::SchemaValidator::OpenAPI3::StringParamsCoercer.
+          new(validator_option).
+          coerce_parameter_object(query_hash, path_parameters)
     end
 
-    def coerce_path_parameter_object(name, value)
-      path_parameter_object = path_parameters[name]
-      return [nil, false] unless path_parameter_object
-
-      coerce_value(value, path_parameter_object)
+    def coerce_query_parameter(query_hash, validator_option)
+      Committee::SchemaValidator::OpenAPI3::StringParamsCoercer.
+          new(validator_option).
+          coerce_parameter_object(query_hash, query_parameters)
     end
 
     def validate_request_params(params)
@@ -196,41 +194,6 @@ module Committee
 
     def path_parameters
       @path_parameters ||= oas_parser_endpoint.path_parameters.map{ |parameter| [parameter.name, parameter] }.to_h
-    end
-
-    def coerce_value(value, query_parameter_object)
-      return [nil, false] unless value
-
-      schema = query_parameter_object.schema
-      return [value, false] unless schema
-
-      case schema["type"]
-      when "integer"
-        begin
-          return Integer(value), true
-        rescue ArgumentError => e
-          raise e unless e.message =~ /invalid value for Integer/
-        end
-      when "boolean"
-        if value == "true" || value == "1"
-          return true, true
-        end
-        if value == "false" || value == "0"
-          return false, true
-        end
-      when "number"
-        begin
-          return Float(value), true
-        rescue ArgumentError => e
-          raise e unless e.message =~ /invalid value for Float/
-        end
-      else
-        # nothing to do
-      end
-
-      return [nil, true] if schema["nullable"] && value == ""
-
-      [nil, false]
     end
 
     # check responses object
